+++ src/lib/route-data/index.ts
@@ -1,3 +1,3 @@
 export { LabelFunction } from './route-data-function-types';
 export { RouteDataInterface } from './route-data.interface';
-export { RouteDataRegistry } from './route-data.registry';
+export { RouteDataRegistry } from './route-data-registry';
